Make the entity id max greater then ids datomic generates
This allows centriq-web frontend, which uses those datomic ids, to continue to
work without any modification. Here is a slack conversation between me and the
author that lead to this idea:

https://clojurians.slack.com/archives/C07V8N22C/p1662565926692069

Though the important part is simply that he suggests increasing the number(s) is
ok. I choose to only increase the emax, as It's unclear why the transaction id min
or max would need to be changed.

NOTE i set the emax to the hexed version of max safe integer in javascript. aka
> Number.MAX_SAFE_INTEGER
<- 9007199254740991

TODO i also haven't verified the max size integer datomic might pass.

(I assume using hex is faster)

---------------------------------------------------------------------------------

The alternative solutions, to increasing the emax, involve syncing via a
server/id (see datasync repo for details on that implementation), as well as
updates a unknown about of code that relies on ids being exactly as they are now!

Or you could sync using a uuid, which if persisted and added to datomic, would
have the advantage of being a universal way of identifying entities, which could
be useful if were commenting cross databases. Given the FE centriq app is more
accurately a "cache" then a new database, this isn't as good as "easy" as using
the ids. If it's "simpler" is probably a matter of implementation details, as
the need for this fork/change suggests the db/id is for the database and isn't
necessarily ok crossing borders!